Catagory section created
When user clicks on a tag from the dropdown section of the navbar,
posts with corresponding catagory will show up. Uses a get request to
accomplish this.
Some temporary styling was put in place as well.

NOTE: Right now filtering logic is within the view, I'd really like to
refactor this at some point.